requirements.txt
retry~=0.9.2
SQLAlchemy~=1.4.0
tushare~=1.2.84
PyYAML~=5.3.1
pandas~=1.4.2
psycopg2